refactor(frontend): reorganize settings screen folders
