+++ cmd/gomander/frontend/src/screens/SettingsScreen/schemas/projectSettingsSchema.ts
