+++ cmd/gomander/frontend/src/screens/SettingsScreen/schemas/userSettingsSchema.ts
